Add files via upload
Problem Statement for MinProduct
You are given an array amount with exactly 10 elements, where the i-th element denotes the number of copies of the digit i you have.

You would like to construct two nonnegative integers A and B. The base-10 representation of A must have exactly blank1 digits. The base-10 representation of B must have exactly blank2 digits. It is allowed for A and B to contain leading zeros. In addition, the total number of appearances of digit i in the numbers A and B must be at most amount[i]. It's guaranteed that the sum of amount is at least blank1 + blank2.

Among all valid pairs (A, B), find the one for which A * B is minimal. Return the smallest possible value of the product A * B.

Problem statement for DistinctGridEasy
You are given s n,k and a grid.

Cat Noku has constructed a 2d matrix of integers with dimensions n by n. Each element of the matrix is between 0 and k-1, inclusive. You are given the 2d matrix in a special format, in the grid. grid has e